requirements.txt
anyio==3.7.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.97.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
opensearch==0.9.2
opensearch-py==2.2.0
psycopg2-binary==2.9.6
pydantic==1.10.9
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.16
starlette==0.27.0
typing_extensions==4.6.3
urllib3==1.26.16
uvicorn==0.22.0
